Перенос Handler и соответственно всей связанной с ним логики обратно в активити
