+++ app/src/main/java/com/practicum/playlistmaker/player/domain/api/InteractorPlayer.kt
@@ -3,7 +3,7 @@ package com.practicum.playlistmaker.player.domain.api
 import com.practicum.playlistmaker.player.domain.model.PlayerState
 
 interface InteractorPlayer {
-    fun prepare(url: String, onComplete: () -> Unit)
+    fun prepare(url: String)
     fun play()
     fun pause()
     fun release()
