+++ app/src/main/java/com/practicum/playlistmaker/player/domain/api/RepositoryPlayer.kt
@@ -3,7 +3,7 @@ package com.practicum.playlistmaker.player.domain.api
 import com.practicum.playlistmaker.player.domain.model.PlayerState
 
 interface RepositoryPlayer {
-    fun prepare(url: String, onComplete: () -> Unit)
+    fun prepare(url: String)
     fun play()
     fun pause()
     fun release()
