+++ app/src/main/java/com/practicum/playlistmaker/player/domain/interactor/InteractorPlayerImpl.kt
@@ -7,8 +7,8 @@ import com.practicum.playlistmaker.player.domain.model.PlayerState
 class InteractorPlayerImpl(
     private val repository: RepositoryPlayer
 ): InteractorPlayer {
-    override fun prepare(url: String, onComplete: () -> Unit) {
-        repository.prepare(url, onComplete)
+    override fun prepare(url: String) {
+        repository.prepare(url)
     }
 
     override fun play() {
