+++ app/src/main/java/com/practicum/playlistmaker/search/ui/activity/SearchActivity.kt
@@ -5,6 +5,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import android.os.SystemClock
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.View
@@ -22,10 +23,7 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.practicum.playlistmaker.util.Creator
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.search.domain.consumer.Consumer
-import com.practicum.playlistmaker.search.domain.consumer.ConsumerData
 import com.practicum.playlistmaker.search.domain.model.Track
 import com.practicum.playlistmaker.player.ui.activity.PlayerActivity
 import com.practicum.playlistmaker.search.ui.model.SearchHistoryState
@@ -36,6 +34,9 @@ import java.io.Serializable
 class SearchActivity : AppCompatActivity() {
     companion object{
         const val PLAYER_INTENT_KEY = "player_intent_key"
+        private const val REQUECT_KEY = "REQUEST_KEY"
+        private const val CLICK_DEBOUNCE_DELAY = 1000L
+        private const val SEARCH_DEBOUNCE_DELAY = 2000L
     }
 
     val viewModel by lazy { ViewModelProvider(this)[SearchViewModel::class.java] }
@@ -53,6 +54,28 @@ class SearchActivity : AppCompatActivity() {
     private var searchResultsAdapterList = mutableListOf<Track>()
     private var historyAdapterList = mutableListOf<Track>()
 
+    private val handler = Handler(Looper.getMainLooper())
+
+    fun searchDebounce(searchText: String){
+        handler.removeCallbacksAndMessages(REQUECT_KEY)
+        val searchRunnable = Runnable { viewModel.loadData(searchText) }
+        val postTime = SystemClock.uptimeMillis() + SEARCH_DEBOUNCE_DELAY
+        handler.postAtTime(
+            searchRunnable,
+            REQUECT_KEY,
+            postTime,
+        )
+    }
+
+    private var isClickAllowed = true
+    fun clickDebounce() : Boolean {
+        val current = isClickAllowed
+        if (isClickAllowed) {
+            isClickAllowed = false
+            handler.postDelayed({ isClickAllowed = true }, CLICK_DEBOUNCE_DELAY)
+        }
+        return current
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -87,7 +110,7 @@ class SearchActivity : AppCompatActivity() {
         }
 
         err_btn_refrech.setOnClickListener{
-            viewModel.searchDebounce(searchText)
+            searchDebounce(searchText)
         }
 
         btn_clear.setOnClickListener{
@@ -128,13 +151,13 @@ class SearchActivity : AppCompatActivity() {
 
             override fun afterTextChanged(s: Editable?) {
                 searchText = s.toString()
-                viewModel.searchDebounce(searchText)
+                searchDebounce(searchText)
             }
         }
         search_bar.addTextChangedListener(searchTextWatcher)
 
         searchResultsAdapter = TrackAdapter(searchResultsAdapterList) { track ->
-            if (viewModel.clickDebounce()) {
+            if (clickDebounce()) {
                 viewModel.add(track)
                 val playerIntent = Intent(this, PlayerActivity::class.java)
                 startActivity(playerIntent.putExtra(PLAYER_INTENT_KEY, track as Serializable))
@@ -144,7 +167,7 @@ class SearchActivity : AppCompatActivity() {
         searchRecycleView.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
 
         historyAdapter = TrackAdapter(historyAdapterList) { track ->
-            if (viewModel.clickDebounce()) {
+            if (clickDebounce()) {
                 val playerIntent = Intent(this, PlayerActivity::class.java)
                 startActivity(playerIntent.putExtra(PLAYER_INTENT_KEY, track as Serializable))
             }
